dont install magick if its already exist

+++ api_packages/recipes/default.rb
@@ -4,19 +4,26 @@ bash "Install API packages" do
     sudo apt-get install -y libav-tools libmagickcore-dev libmagickwand-dev libmagic-dev libpq-dev
     sudo apt-get install -y build-essential checkinstall && apt-get build-dep imagemagick
 
-    cd /home/ubuntu
-    rm -rf ImageMagick*
-    mkdir ImageMagick
-    sudo wget http://www.imagemagick.org/download/ImageMagick.tar.gz
-    sudo tar xzvf ImageMagick.tar.gz -C /home/ubuntu/ImageMagick --strip-components=1
+    MAGICK_PATH=$(which magick)
 
-    cd ImageMagick
-    sudo ./configure
-    sudo make clean
-    sudo make
-    sudo checkinstall
-    sudo sh -c "ldconfig /usr/local/lib"
-    magick -version
+    if [[ $MAGICK_PATH = ""  ]]; then
+        cd /home/ubuntu
+        rm -rf ImageMagick*
+        mkdir ImageMagick
+        sudo wget http://www.imagemagick.org/download/ImageMagick.tar.gz
+        sudo tar xzvf ImageMagick.tar.gz -C /home/ubuntu/ImageMagick --strip-components=1
+
+        cd ImageMagick
+        sudo ./configure
+        sudo make clean
+        sudo make
+        sudo checkinstall
+        sudo sh -c "ldconfig /usr/local/lib"
+        magick -version
+    else
+        echo magic exists
+        echo $MAGICK_PATH
+    fi
 
   EOF
   user "ubuntu"
